:recycle: [Refactor] 프로필 업데이트 하는 부분 리팩토링 #main
