+++ src/main.js
@@ -1,5 +1,10 @@
 import { renderPage } from "./router.js";
-import { handleLinkClick, handleLogout, handleLogin } from "./eventHandlers.js";
+import {
+  handleLinkClick,
+  handleLogout,
+  handleLogin,
+  handleUpdateProfile,
+} from "./eventHandlers.js";
 
 document.addEventListener("click", (event) => {
   handleLinkClick(event);
@@ -8,15 +13,7 @@ document.addEventListener("click", (event) => {
 
 document.addEventListener("submit", (event) => {
   handleLogin(event);
-
-  if (event.target.id === "profile-form") {
-    event.preventDefault();
-    const username = document.querySelector("#username").value;
-    const email = document.querySelector("#email").value;
-    const bio = document.querySelector("#bio").value;
-    localStorage.setItem("user", JSON.stringify({ username, email, bio }));
-    alert("프로필이 업데이트되었습니다.");
-  }
+  handleUpdateProfile(event);
 });
 
 window.addEventListener("popstate", () => {
